Moved new tests to WolvenKit.UnitTests
