+++ Tests/WolvenKit.UnitTests/FNV1A64Tests.cs
@@ -8,6 +8,9 @@ namespace WolvenKit.UnitTests
     [TestClass]
     public class FNV1A64Tests
     {
+        private const ulong s_helloWorldHash = 14998808953798625955;
+        private const ulong s_semicolonHash = 16982440701406162758;
+        
         [TestMethod]
         [DataRow("", 0xcbf29ce484222325UL)]
         [DataRow("a", 0xaf63dc4c8601ec8cUL)]
@@ -92,6 +95,38 @@ namespace WolvenKit.UnitTests
         [DataRow("hello", 0xa9bc8acca21f39b1UL)]
         [DataRow("foobar", 0x34531ca7168b8f38UL)]
         [DataRow("feedfacedeadbeef", 0xa7a4c9f3edebf0d8UL)]
-        public void TestFNV1a64_NullEnded(string test, ulong result) => Assert.AreEqual(FNV1A64HashAlgorithm.HashString(test, Encoding.ASCII, true), result);
+        public void TestFNV1a64_NullEnded(string test, ulong result) =>
+            Assert.AreEqual(FNV1A64HashAlgorithm.HashString(test, Encoding.ASCII, true), result);
+
+        [TestMethod]
+        [DataRow("$/hello/world/12345")]
+        [DataRow("$/#hello/world/12345")]
+        public void Hasher_Should_Handle_Simple_Cases(string source)
+        {
+            var hash = FNV1A64HashAlgorithm.HashStringWithoutAliases(source);
+            Assert.AreEqual(hash, s_helloWorldHash);
+        }
+
+        [TestMethod]
+        [DataRow("$/test/important/123/456")]
+        [DataRow("$/test/important;#alias/123/456")]
+        public void Hasher_Should_Handle_Semicolon(string source)
+        {
+            var hash = FNV1A64HashAlgorithm.HashStringWithoutAliases(source);
+            Assert.AreEqual(hash, s_semicolonHash);
+        }
+
+        [TestMethod]
+        [DataRow("$/03_night_city/#c_city_center/#c_city_center_air_traffic", 13701749328974838295)]
+        [DataRow("$/03_night_city/#c_city_center/#c_city_center_air_traffic/{air_traffic_av_line_mid_8_avs}_020_prefabQ5RFZPI/{air_traffic_mid_}_002_prefabSDVGHLQ", 5910196056938917739ul)]
+        [DataRow("$/03_night_city/#c_city_center/#c_city_center_air_traffic/{air_traffic_av_line_mid_8_avs}_020_prefabQ5RFZPI/speedSpline", 1868123494098026698ul)]
+        [DataRow("$/03_night_city/#c_city_center/corpo_plaza/#loc_arasaka_tower/#loc_arasaka_tower_env/interior_prefab5W2RVJA/q113_saburos_office_prefab3J25ZSA/{loc_q113_arasaka_rooftop_devices}_prefabWH2FUBQ/{q113_arasaka_offices_main_door_new}_prefab7RDD2FQ/{double_door_sliding_l}_prefabV6UV7YQ;#q113_arasaka_offices_main_door_new", 4886409109574024977ul)]
+        [DataRow("$/03_night_city/#c_city_center/corpo_plaza/#loc_arasaka_tower/#loc_arasaka_tower_env/interior_prefab5W2RVJA/q113_saburos_office_prefab3J25ZSA/{loc_q113_arasaka_rooftop_devices}_prefabWH2FUBQ/lift_3_floors_prefab4PDNO3Y/{lift_door_}3_prefab7Y7O7ZI_005;#q113_06_saburo_office_hallway_elevator", 12677911292092322200)]
+        [DataRow("$/03_night_city/#c_city_center/corpo_plaza/#loc_arasaka_tower/#loc_arasaka_tower_env/interior_prefab5W2RVJA/q113_saburos_office_prefab3J25ZSA/{loc_q113_arasaka_rooftop_devices}_prefabWH2FUBQ/single_door_5_prefabP56IZMY/door1;#q113_av_hangar_door", 987549654063553389ul)]
+        public void Hasher_Should_Handle_Real_Life_Examples(string source, ulong expected)
+        {
+            var hash = FNV1A64HashAlgorithm.HashStringWithoutAliases(source);
+            Assert.AreEqual(hash, expected);
+        }
     }
 }
